requirements.txt
Pillow~=9.4.0
PyQt5~=5.15.4
wifi-qrcode-generator~=0.1